backend/pdf_tron_app/serializers.py

# import serializer from rest_framework
from rest_framework import serializers
from .models import FileUpload


class FileUploadSerializer(serializers.ModelSerializer):
    class Meta:
        model = FileUpload
        fields = '__all__'